refactor: split neovim agent terminal modules
